Module finchers_headers::header

The definition of header types.

This module contains some re-exported types from hyperx crate.

Structs

Accept
Accept header, defined in RFC7231

AcceptCharset
Accept-Charset header, defined in RFC7231

AcceptEncoding
Accept-Encoding header, defined in RFC7231

AcceptLanguage
Accept-Language header, defined in RFC7231

AcceptRanges
Accept-Ranges header, defined in RFC7233

Connection
Connection header, defined in RFC7230

ContentDisposition
A Content-Disposition header, (re)defined in RFC6266.

ContentEncoding
Content-Encoding header, defined in RFC7231

ContentLanguage
Content-Language header, defined in RFC7231

ContentLength
Content-Length header, defined in RFC7230

ContentType
Content-Type header, defined in RFC7231

Date
Date header, defined in RFC7231

EntityTag
An entity tag, defined in RFC7232

Host
The Host header.

HttpDate
A timestamp with HTTP formatting and parsing

IfModifiedSince
If-Modified-Since header, defined in RFC7232

IfUnmodifiedSince
If-Unmodified-Since header, defined in RFC7232

LanguageTag
A language tag as described in BCP47.

Mime
A parsed mime or media type.

Origin
The Origin header.

Quality
Represents a quality used in quality values.

QualityItem
Represents an item with a quality value as defined in RFC7231.

Referer
Referer header, defined in RFC7231

Te
TE header, defined in RFC7230

UserAgent
User-Agent header, defined in RFC7231

Warning
Warning header, defined in RFC7234

Enums

Charset
A Mime charset.

ConnectionOption
Values that can be in the Connection header.

DispositionParam
A parameter to the disposition type.

DispositionType
The implied disposition of the content of the HTTP body.

Encoding
A value to represent an encoding used in Transfer-Encoding or Accept-Encoding header.

Expect
The Expect header.

IfMatch
If-Match header, defined in RFC7232

IfNoneMatch
If-None-Match header, defined in RFC7232

IfRange
If-Range header, defined in RFC7233

RangeUnit
Range Units, described in RFC7233

Traits

TypedHeader
A trait representing a pair of header field and value.

Type Definitions

QualifiedVec
Type alias represents a Vec which holds qualified items.