Trait finchers_headers::header::TypedHeader

pub trait TypedHeader {
    type Output;
    type Error: Into<Error>;

    const NAME: HeaderName;

    fn parse_header(value: &HeaderValue) -> Result<Self::Output, Self::Error>;
}

A trait representing a pair of header field and value.

Associated Types

type Output

The value type of header field which will be returned from parse_header.

type Error: Into<Error>

The error type which will be returned from parse_header.

Associated Constants

const NAME: HeaderName

The name of header field associated with this type.

Required Methods

fn parse_header(value: &HeaderValue) -> Result<Self::Output, Self::Error>

Parse a header from a raw header value.